+++ hal/CHANGELOG.md
@@ -1,7 +1,9 @@
-v0.x
+v0.3
 ---
 - Updates to GPIO header.
 > Remove alternate function callback, replace with singular value.
+- Addition of serial header.
+> Add functions to set-up/teardown, transmit and receive, callback.
 
 v0.2
 ---